qca-wifi: Use rssi_chain value in uint8_t format for EWMA averaging
The value of rssi_chain can vary between -128 to 127.

For peer extended rate stats, average rssi_chain is computed using
EWMA averaging mechanism which expects the variable data type to be
of unsigned long format.

In case of negative rssi_chain value, avg value turns out to be a
very big value which does not hold any significance.

In order to fix this, use rssi_chain in uint8_t format for average
computation and in the flush output, print the average value in
int8_t format.

CRs-Fixed: 3177194

+++ tools/linux/peerstats.c
@@ -213,8 +213,8 @@ static void dp_peer_rx_rate_stats_print(uint8_t *peer_mac,
 			printf(" %10u |", tmp_rx_stats->rate);
 			for (chain = 0; chain < max_chain; chain++) {
 				for (bw = 0; bw < max_bw; bw++) {
-					printf(" %10lu |",
-					tmp_rx_stats->avg_rssi_ant[chain][bw]);
+					printf(" %10d |",
+					(int8_t)tmp_rx_stats->avg_rssi_ant[chain][bw]);
 				}
 				printf("            \n\t     ");
 			}
